requirements.txt
kagglehub==0.3.12
matplotlib==3.10.3
numpy==2.3.1
Pillow==11.3.0
scikit_learn==1.7.0
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.0
ipython==8.12.3
onnx==1.18.0

ezkl==22.0.1
nest-asyncio==1.6.0
pytictoc==1.5.3